weapon/requirements.txt
ultralytics>=8.3.0
opencv-python>=4.8.0
requests>=2.31.0
PyYAML>=6.0
python-telegram-bot>=20.0
numpy>=1.24.0,<2.0.0
Pillow>=10.0.0
python-dotenv>=1.0.0
torch>=2.0.0
torchvision>=0.15.0
Flask>=2.3.0
Werkzeug>=2.3.0 